Cleaning up XML and repository positions in code
