+++ src/Syringe.Core/Repositories/Xml/Reader/ITestFileReader.cs
@@ -1,0 +1,10 @@
+﻿using System.IO;
+using Syringe.Core.Tests;
+
+namespace Syringe.Core.Repositories.XML.Reader
+{
+	public interface ITestFileReader
+    {
+		TestFile Read(TextReader textReader);
+    }
+}
