+++ src/Syringe.Core/Repositories/Xml/Writer/ITestFileWriter.cs
@@ -1,6 +1,6 @@
 using Syringe.Core.Tests;
 
-namespace Syringe.Core.Xml.Writer
+namespace Syringe.Core.Repositories.XML.Writer
 {
 	public interface ITestFileWriter
 	{
