--- src/Syringe.Core/Xml/Reader/ITestFileReader.cs
@@ -1,10 +1,0 @@
-﻿using System.IO;
-using Syringe.Core.Tests;
-
-namespace Syringe.Core.Xml.Reader
-{
-	public interface ITestFileReader
-    {
-		TestFile Read(TextReader textReader);
-    }
-}
