+++ src/Syringe.Service/DependencyResolution/DefaultRegistry.cs
@@ -28,8 +28,8 @@ using Syringe.Core.IO;
 using Syringe.Core.Repositories;
 using Syringe.Core.Repositories.MongoDB;
 using Syringe.Core.Repositories.XML;
-using Syringe.Core.Xml.Reader;
-using Syringe.Core.Xml.Writer;
+using Syringe.Core.Repositories.XML.Reader;
+using Syringe.Core.Repositories.XML.Writer;
 using Syringe.Service.Api.Hubs;
 using Syringe.Service.Parallel;
 using WebApiContrib.IoC.StructureMap;
