+++ src/Syringe.Service/Parallel/ParallelTestFileQueue.cs
@@ -9,11 +9,11 @@ using RestSharp;
 using Syringe.Core.Configuration;
 using Syringe.Core.Http;
 using Syringe.Core.Repositories;
+using Syringe.Core.Repositories.XML.Reader;
 using Syringe.Core.Runner;
 using Syringe.Core.Tasks;
 using Syringe.Core.Tests;
 using Syringe.Core.Tests.Results;
-using Syringe.Core.Xml.Reader;
 
 namespace Syringe.Service.Parallel
 {
