+++ src/Syringe.Tests/Unit/Repositories/Xml/Reader/XmlExamples/single-test.xml
@@ -1,0 +1,43 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<tests>
+	<test
+        id="1"
+        description="description"
+        url="http://myserver"
+        method="post"
+        expectedhttpstatuscode="404"
+       
+        logrequest="false"
+        logresponse="false"
+	    >
+
+		<headers>
+			<header name="mykey">12345</header>
+			<header name="bar">foo</header>
+			<header name="emptyvalue"></header>
+			<header name="Cookie">referer=harrispilton.com</header>
+		</headers>
+
+		<postbody>
+			<![CDATA[
+            username=corey&password=welcome
+            ]]>
+		</postbody>
+
+		<capturedvariables>
+			<variable name="1">parse 1</variable>
+			<variable name="2">parse 11</variable>
+			<variable name="3">parse 99</variable>
+		</capturedvariables>
+
+		<assertions>
+			<assertion type="positive" description="1" method="regex">positive 1</assertion>
+			<assertion type="PosItiVe" description="2">positive 22</assertion>
+			<assertion type="positive" description="3">positive 99</assertion>
+
+			<assertion type="negative" description="1" method="csquery">negative 1</assertion>
+			<assertion type="NEGaTiVE" description="2">negative 6</assertion>
+			<assertion type="negative" description="3">negative 66</assertion>
+		</assertions>
+	</test>
+</tests>
