+++ src/Syringe.Tests/Unit/Repositories/Xml/Writer/XmlExamples/variables.xml
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<tests>
+	<variables>
+		<variable name="name1" environment="env1">value1</variable>
+		<variable name="name2" environment="env2">value2</variable>
+	</variables>
+</tests>
