--- src/Syringe.Tests/Unit/Xml/XmlExamples/Reader/full.xml
@@ -1,63 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<tests>
-	<variables>
-    <variable name="LOGIN_URL" environment="DevTeam1">http://myserver/login.php</variable>
-    <variable name="LOGIN_URL" environment="DevTeam2">http://myserver2/login.php</variable>
-		<variable name="LOGIN1" environment="DevTeam2">bob</variable>
-		<variable name="PASSWD1" environment="DevTeam1">sponge</variable>
-		<variable name="SUCCESSFULL_TEST_TEXT" environment="DevTeam2">Welcome Bob</variable>
-	</variables>
-
-	<test
-		id="1"
-		description="description 1"
-		method="post"
-		url="http://myserver/test/login.jsp"
-	>
-		<postbody>
-			<![CDATA[username=corey&password=welcome]]>
-		</postbody>
-
-		<assertions>
-			<verify type="positive" description="1" method="regex">verify this string exists</verify>
-			<verify type="negative" description="1" method="regex">verify this string does not exist</verify>
-		</assertions>
-		
-		<capturedvariables>
-			<variable name="test">/d</variable>
-		</capturedvariables>
-	</test>
-
-	<test
-		id="300"
-		description="description 300"
-		method="get"
-		url="http://myserver/test/send.jsp?value={TIMESTAMP}"
-	>
-		<assertions>
-			<assertion type="positive" id="1" method="regex">verify this string exists</assertion>
-		</assertions>
-	</test>
-
-	<test
-		id="20"
-		description="description 20"
-		method="get"
-		url="http://myserver/test/send.jsp?value={TIMESTAMP}"
-	>
-		<assertions>
-			<assertion type="positive" description="1" method="regex">verify this string exists</assertion>
-		</assertions>
-	</test>
-
-	<test
-		id="9"
-		description="description 9"
-		method="post"
-		url="http://myserver/test/send.jsp?value={TIMESTAMP}"
-	>
-		<assertions>
-			<assertion type="positive" description="1" method="regex">verify this string exists</assertion>
-		</assertions>
-	</test>
-</tests>
