--- src/Syringe.Tests/Unit/Xml/XmlExamples/Writer/capturedvariables.xml
@@ -1,21 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<tests>
-	<test
-		description=""
-		url=""
-		method=""
-		expectedhttpstatuscode="0"
-	>
-
-		<headers/>
-		<postbody/>
-
-		<capturedvariables>
-            <variable name="1">here is (.*?) regex</variable>
-            <variable name="2">plain text</variable>
-            <variable name="3">This is encoded &lt;test&gt; &amp;.</variable>
-        </capturedvariables>
-
-		<assertions />
-    </test>
-</tests>
