--- src/Syringe.Tests/Unit/Xml/XmlExamples/Writer/headers.xml
@@ -1,19 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<tests>
-    <test
-        description=""
-        url=""
-        method=""
-        expectedhttpstatuscode="0"
-	    >
-
-        <headers>
-            <header name="key1">value1</header>
-            <header name="key2"><![CDATA[some <marvellous> HTML &&&&.]]></header>
-        </headers>
-
-		<postbody/>
-		<capturedvariables />
-		<assertions />
-    </test>
-</tests>
