--- src/Syringe.Tests/Unit/Xml/XmlExamples/Writer/variables.xml
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<tests>
-	<variables>
-		<variable name="name1" environment="env1">value1</variable>
-		<variable name="name2" environment="env2">value2</variable>
-	</variables>
-</tests>
